requirements.txt
click==8.1.7
colorama==0.4.6
Flask==2.2.5
Flask-Cors==4.0.0
Flask-MySQLdb==2.0.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
PyJWT==2.8.0
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
